Start on better math operations
I was tired of having the weird fake operations that were
in place before.

This commit changes that!
I have added support for chaining operations (ie 1 + 2 + 3...)

As of now, only addition and subtraction are supported, and with only a
few value types. Strings are soon to be updated to coincide with the new
operations system.