Large update
Now you can load files WITHOUT the "--file" argument, by simply
providing the filename after "kepler"

Also, in if statements, you can now chain comparisons with "and" and
"or", as well as outside of if statements.

Finally, the basis for infinite loops has been added with "every". You
are now able to set intervals that run every n milliseconds. Breaking
out of these intervals is as easy as creating them. This can be used as
a good basis for regular infinite loops.