Seperate equality checks
Before, doing an "equals" equality check would compare types and return
false if they were not equal.

Now, however, it will attempt to cast the values to be like, then
compare them.

This left a hole for strict equality checks, which have now also been
added. The operator "==" is used for strict equality checks, which will
return false if types are not equal.