Fix constant validation in type check & improve formatting
For some reason, when checking types, an error would be thrown when the
variable was declared as a constant. This was fixed by removing the
constant check from the type checker.

Also, the formatting of the VariableManager and FunctionManager was
improved. There are still some snags to be worked out, but it's is a
good start.