Optimize and add caching
Before, a new Interpreter would be created for every conditional and
function call. This is inefficient and heavy. Now, we can cache the
Interpreter and use it again for the next conditional or function call.

Originally, a Sieve of Eratosthenes up to 10k would take up to 30
seconds to complete. Now, after enabling code optimization and adding
caching, it completes in less than 3 seconds.